+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Based on original code from [NIMH MEG Core Facility].
 
 (c) 2021-2023 Claudio Satriano <satriano@ipgp.fr>
 
-## Unreleased
+## v1.1 - 2023-06-05
 
 - Use ctypes to wrap the modules written in C
 - License changed to GPLv3